requirements.txt
anyio==3.5.0
asgiref==3.4.1
autopep8==1.6.0
bidict==0.22.0
click==8.0.3
fastapi==0.72.0
h11==0.12.0
idna==3.3
pycodestyle==2.8.0
pydantic==1.9.0
python-engineio==4.3.2
python-socketio==5.6.0
sniffio==1.2.0
starlette==0.17.1
toml==0.10.2
typing_extensions==4.0.1
uvicorn==0.17.0
